renaming react component to match export map
